requirements.txt
langchain-community
openai
faiss-cpu
chromadb
PyMuPDF
tiktoken
